Switch prime field class from i64 to i128
We might want to allow any integers to be used here such that
PrimeFieldElement -> PrimeFieldElement<T: Num>. But for now, we just use
128-bit numbers. I should look into whether the compiler generates
efficient code, when type parameters are used. I suspect that it *does*
since the types can be determined compile-time and the functions can
thus be adjusted to the known bit lengths.

+++ benches/ntt.rs
@@ -8,7 +8,7 @@ fn generate_ntt_input<'a>(
     let size: usize = 2usize.pow(log2_of_size as u32);
     let mut output: Vec<PrimeFieldElement<'a>> = Vec::with_capacity(size);
     for _ in 0..size {
-        let rand = rand::random::<u32>() as i64 % prime.q;
+        let rand = rand::random::<u32>() as i128 % prime.q;
         output.push(PrimeFieldElement::new(rand, prime));
     }
 
@@ -16,11 +16,11 @@ fn generate_ntt_input<'a>(
 }
 
 fn ntt_forward(c: &mut Criterion) {
-    static PRIME: i64 = 167772161; // 5 * 2^25 + !
+    static PRIME: i128 = 167772161; // 5 * 2^25 + !
     let prime_field: PrimeField = PrimeField::new(PRIME);
     let mut group = c.benchmark_group("ntt_forward");
     for log2_of_size in [18usize, 19, 20, 21, 22].iter() {
-        let size = 2i64.pow(*log2_of_size as u32);
+        let size = 2i128.pow(*log2_of_size as u32);
         let unity_root = prime_field.get_primitive_root_of_unity(size);
         group.throughput(Throughput::Bytes(*log2_of_size as u64));
         group
